refactor: Move EntityById to transit package.
