+++ src/main/java/org/opentripplanner/netex/mapping/NoticeMapper.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.netex.mapping;
 
-import org.opentripplanner.model.impl.EntityById;
 import org.opentripplanner.netex.mapping.support.FeedScopedIdFactory;
 import org.opentripplanner.transit.model.basic.Notice;
+import org.opentripplanner.transit.model.framework.EntityById;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
 /**
